requirements.txt
https://github.com/Rapptz/discord.py/archive/rewrite.zip#egg=discord.py[voice]
apng
arrow
GitPython
html2text
lxml
groupyapi
Pybooru
asyncjisho
simplejson
monthdelta
embypy>=0.3.2.0
matplotlib
networkx
Pillow
